pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "logd-predictor"
version = "0.1.0"
description = "ChEMBL logD data pipeline and baseline model training"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "chembl-downloader>=0.5.2",
    "deepchem>=2.8.0",
    "hydra-core>=1.3.0",
    "hydra-optuna-sweeper>=1.2.0",
    "hydraflow>=0.6.0",
    "mlflow>=2.10.0",
    "optuna>=2.10.0,<3.0.0",
    "ipython>=9.12.0",
    "pandas>=2.2.0",
    "pip>=26.0.1",
    "rich>=13.7.0",
    "rdkit>=2026.3.1",
    "scikit-learn>=1.4.0",
    "torch>=2.11.0",
    "lightning>=2.6.1",
    "torchmetrics>=1.9.0",
    "joblib>=1.5.3",
    "ruff>=0.15.11",
    "autopep8>=2.3.2",
]

[project.scripts]
chembl-scraper = "logd_predictor.chembl_scraper:main"
chembl-logd-pipeline = "logd_predictor.pipeline:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["logd_predictor*"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
